+++ tests/roots/test-ext-autodoc/target/reimport/_private.py
@@ -1,0 +1,4 @@
+from typing import TypeVar
+
+T = TypeVar('T', bound=int | str)
+"""A reimported TypeVar"""