+++ CHANGES.rst
@@ -127,6 +127,7 @@ Bugs fixed
 * #12975: Avoid rendering a trailing comma in C and C++ multi-line signatures.
 * #13178: autodoc: Fix resolution for ``pathlib`` types.
   Patch by Adam Turner.
+* #13276: autodoc: Allow TypeVars to be reimported from other modules
 
 Testing
 -------